# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: HaoHaoAccounting API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /transactions:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [transactions]
      operationId: getTransactions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: categoryId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: accountId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Transaction list
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Link:
              $ref: '#/components/headers/Link'
            X-Total-Count:
              $ref: '#/components/headers/TotalCount'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
    post:
      tags: [transactions]
      operationId: postTransactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Created transaction
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /transactions/{id}:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    put:
      tags: [transactions]
      operationId: putTransactionsById
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Updated transaction
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [transactions]
      operationId: deleteTransactionsById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
  /io/import/preview:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [dataio]
      operationId: postIoImportPreview
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportFileRequest'
      responses:
        '200':
          description: Import preview
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /io/import:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [dataio]
      operationId: postIoImport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportFileRequest'
      responses:
        '200':
          description: Import result
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /io/import/jobs:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [dataio]
      operationId: getIoImportJobs
      responses:
        '200':
          description: Import job history
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImportJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
    post:
      tags: [dataio]
      operationId: postIoImportJobs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportFileRequest'
      responses:
        '202':
          description: Queued import job
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /io/import/jobs/{id}:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [dataio]
      operationId: getIoImportJobsById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Import job status
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /io/import/text/preview:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [dataio]
      operationId: postIoImportTextPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportTextRequest'
      responses:
        '200':
          description: Text import preview
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /io/import/text:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [dataio]
      operationId: postIoImportText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportTextRequest'
      responses:
        '200':
          description: Text import result
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /io/export:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [dataio]
      operationId: getIoExport
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, xlsx]
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Exported ledger file
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Content-Disposition:
              $ref: '#/components/headers/ContentDisposition'
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional request correlation id. The API returns this value when valid, otherwise it generates one.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
  headers:
    RequestID:
      description: Request correlation id for tracing logs and API calls.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
    Link:
      description: RFC 8288 pagination links for first, previous, next, and last pages when available.
      schema:
        type: string
    TotalCount:
      description: Total number of records matching the query before pagination.
      schema:
        type: integer
        minimum: 0
    ContentDisposition:
      description: Attachment disposition for downloaded files. Includes both filename and RFC 5987 filename* parameters.
      schema:
        type: string
    Vary:
      description: Request headers that influenced response representation selection.
      schema:
        type: string
    Allow:
      description: HTTP methods supported by the target route.
      schema:
        type: string
    WWWAuthenticate:
      description: Authentication challenge for protected resources. Invalid bearer tokens include error="invalid_token" per RFC 6750.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request or validation error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired authentication token.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Requested resource was not found.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: HTTP method is not supported for this route.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        Allow:
          $ref: '#/components/headers/Allow'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PayloadTooLarge:
      description: Request body is too large.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: Request Content-Type is not supported by this operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotAcceptable:
      description: Requested response media type is not available for this operation.
      headers:
        Vary:
          $ref: '#/components/headers/Vary'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: Request processing exceeded the server timeout budget.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Ok:
      description: OK response
      headers:
        Vary:
          $ref: '#/components/headers/Vary'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'
  schemas:
    TransactionType:
      type: string
      enum: [income, expense]
...
    Account:
      type: object
      required: [id, userId, name, type, balance]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
        balance:
          type: number
...
    Category:
      type: object
      required: [id, name, type, isSystem]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        isSystem:
          type: boolean
...
    Transaction:
      type: object
      required: [id, userId, type, amount, categoryId, accountId, note, tags, source, occurredAt, category, account]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
        categoryId:
          type: integer
        accountId:
          type: integer
        note:
          type: string
        tags:
          type: string
        source:
          type: string
        occurredAt:
          type: string
          format: date-time
        category:
          $ref: '#/components/schemas/Category'
        account:
          $ref: '#/components/schemas/Account'
    TransactionRequest:
      type: object
      additionalProperties: false
      required: [type, amount, categoryId, accountId, note]
      properties:
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          exclusiveMinimum: 0
        categoryId:
          type: integer
          minimum: 1
        accountId:
          type: integer
          minimum: 1
        note:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
        occurredAt:
          type: string
          format: date-time
        source:
          type: string
    TransactionListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        total:
          type: integer
          minimum: 0
...
    ImportFileRequest:
      type: object
      additionalProperties: false
      description: CSV/XLSX import file. The first row must start with occurred_at,type,amount,category,account,note,tags; additional columns are ignored.
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: CSV or XLSX file. The first row must start with occurred_at,type,amount,category,account,note,tags.
        skipDuplicates:
          type: boolean
    ImportTextRequest:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        filename:
          type: string
        content:
          type: string
          description: CSV text. The first row must start with occurred_at,type,amount,category,account,note,tags.
          minLength: 1
          maxLength: 5242880
        skipDuplicates:
          type: boolean
    ImportPreviewRow:
      type: object
      required: [line, occurredAt, type, amount, category, account, note, tags, valid, duplicate]
      properties:
        line:
          type: integer
        occurredAt:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
        category:
          type: string
        account:
          type: string
        note:
          type: string
        tags:
          type: string
        valid:
          type: boolean
        duplicate:
          type: boolean
        error:
          type: string
        duplicateReason:
          type: string
    ImportPreview:
      type: object
      required: [filename, size, totalRows, validRows, failedRows, duplicateRows, maxRows, maxFileBytes, truncated, rows]
      properties:
        filename:
          type: string
        size:
          type: integer
        totalRows:
          type: integer
        validRows:
          type: integer
        failedRows:
          type: integer
        duplicateRows:
          type: integer
        maxRows:
          type: integer
        maxFileBytes:
          type: integer
        truncated:
          type: boolean
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ImportPreviewRow'
    ImportResult:
      type: object
      required: [total, success, failed, skipped, errors]
      properties:
        total:
          type: integer
        success:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
    ImportJob:
      type: object
      required: [id, filename, status, total, success, failed, skipped, errors, createdAt, updatedAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        total:
          type: integer
        success:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [bad_request, invalid_request, unauthorized, forbidden, not_found, method_not_allowed, rate_limited, payload_too_large, unsupported_media_type, not_acceptable, request_timeout, client_closed_request, internal_error]
        requestId:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean